Use setuptools_scm
CI is failing because `requirements.txt` is referenced from `setup.py`
but it not included in the wheel.  The simplest way to fix this is to
adopt setuptools_scm, as I have done here.  An alternative way would
be to provide a `MANIFEST.in` file that includes `requirements.txt`.

Everything works locally because a wheel is not built locally to save
a bit of time if developing rapidly.  For instance, Qiskit Terra's tox
config does not built a wheel for this reason.  I thought I had found
a decent middle ground by builing the wheel _only_ in CI, but it is
perhaps inconvenient (to say the least) that you were experiencing
the failure only there.